feat(nuxt-base): abstract setting cookie session
